+++ packages/config/src/log/logger.js
@@ -1,7 +1,8 @@
 import figures from 'figures'
 
-import { serializeObject } from '../../lib/log/serialize.js'
-
+// FIXME: This error will go away once this file is converted to TypeScript.
+// eslint-disable-next-line n/no-missing-import
+import { serializeObject } from './serialize.js'
 import { THEME } from './theme.js'
 
 export const logsAreBuffered = (logs) => {