+++ packages/config/src/types/integrations.ts
@@ -1,9 +1,0 @@
-export type Integration = {
-  slug: string
-  version?: string
-  has_build?: boolean
-  dev?: {
-    path: string
-  }
-  author?: string
-}